pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-analyzer"
version = "0.1.0"
description = "日本株式市場分析 & おすすめ銘柄選別システム"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.30",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "pandas-ta>=0.3.14b1",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "fugashi>=1.3.0",
    "unidic-lite>=1.0.8",
    "ipadic>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "apscheduler>=3.10.4",
    "rich>=13.7.0",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
